+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Big refactor to make the type decoding and value reading be structured instead of it all being strings
 - Added reading capability for tagged unions (fancy Rust enums)
 - Added colorized output
+- Made it so that transparent types can be added to lessen the clutter in the trace
 
 ## Capture 0.2.0 (03-05-22)